ch(refactor): API endpoints project structure refactored - renamed router main folder [finishes #164558065]
