+++ src/app.js
@@ -2,8 +2,8 @@ import express from 'express';
 import logger from 'morgan';
 import swagger from 'swagger-ui-express';
 import swaggerDocument from './config/swagger.json';
-import userRoute from './api/V1/routes/userRouter';
-import messageRoute from './api/V1/routes/messageRouter';
+import userRoute from './api/V1/router/userRouter';
+import messageRoute from './api/V1/router/messageRouter';
 
 /**
  *
